MFB-1676: group the LIHEAP programs into cross_white_label/liheap/
Seven implementations of one benefit that share no base class: EnergyAssistance,
IlLiheap, KsLieap, MaHeap, NCLieap, TxCeap, WaLiheap. Grouping them by benefit
now means the M2 consolidation is an edit inside one directory.

KS's and WA's specs and tests travel with their calculators, into the family's
specs/ and tests/.
