+++ programs/programs/cross_white_label/liheap/il.py
@@ -1,0 +1,41 @@
+"""IlLiheap."""
+
+from programs.framework.pe_base import PolicyEngineSpmCalulator
+import programs.framework.pe_dependencies as dependency
+
+
+class IlLiheap(PolicyEngineSpmCalulator):
+    """
+    Illinois Low Income Home Energy Assistance Program (LIHEAP).
+
+    Delegates the full eligibility and benefit calculation to PolicyEngine's
+    ``il_liheap`` variable (2026 benefit matrix), rather than hard-coding amounts.
+
+    PolicyEngine computes:
+      - Income eligibility: gross income <= higher of 60% SMI or 200% FPL.
+      - Benefit: an annual matrix amount keyed on fuel type, income bracket, and
+        household size, capped at the household's reported heating expense
+        (``min(matrix_amount, heating)``). PE models every IL household as
+        "All Electric" (or "Cash" when heat is included in rent).
+
+    Required PE input is ``heating_expense_person``: our screener captures heating
+    at the household level, so ``HeatingExpensePersonDependency`` assigns the full
+    amount to the head, which PE aggregates to the SPM unit for the cap. Because
+    the benefit is capped at this expense, a household reporting no heating/cooling
+    expense receives $0.
+
+    Households that already receive IL LIHEAP are flagged and handled by the
+    results layer (via ``already_has``), not in this calculator.
+    """
+
+    program_code = "il_liheap"
+
+    pe_name = "il_liheap"
+    pe_inputs = [
+        dependency.household.IlStateCodeDependency,
+        *dependency.irs_gross_income,
+        dependency.spm.HasHeatingCoolingExpenseDependency,
+        dependency.member.HeatingExpensePersonDependency,
+        dependency.spm.ElectricityExpenseDependency,
+    ]
+    pe_outputs = [dependency.spm.IlLiheap]
